+++ application/core/src/main/java/com/github/pieter_groenendijk/utils/scheduling/LongTermTaskScheduler.java
@@ -1,5 +1,7 @@
 package com.github.pieter_groenendijk.utils.scheduling;
 
+import com.github.pieter_groenendijk.model.Task;
+
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.List;
