+++ application/core/src/main/java/com/github/pieter_groenendijk/utils/scheduling/TaskStorage.java
@@ -1,5 +1,7 @@
 package com.github.pieter_groenendijk.utils.scheduling;
 
+import com.github.pieter_groenendijk.model.Task;
+
 public interface TaskStorage<T extends Task> {
     void store(T task);
 }
